+++ Photo Club Hub/ViewModel/ListReaders/Level2JsonReader.swift
@@ -333,7 +333,8 @@ class Level2JsonReader { // normally running on a background thread
                 featuredImage: featuredImage,
                 featuredImageThumbnail: featuredImage,
                 level3URL: level3URL, // address of portfolio data for this member
-                memberRolesAndStatus: memberRolesAndStatus,
+                memberRolesAndStatus: cleanRoles(memberRolesAndStatus: memberRolesAndStatus,
+                                                 photographer: photographer),
                 fotobondNumber: fotobondNumber,
                 membershipStartDate: membershipStartDate,
                 membershipEndDate: membershipEndDate
@@ -346,6 +347,21 @@ class Level2JsonReader { // normally running on a background thread
         guard jsonOptionals[key].exists() else { return nil }
         guard let string = jsonOptionals[key].string else { return nil }
         return URL(string: string) // returns nil if the string doesn’t represent a valid URL
+    }
+
+    // for deceased members, clear any role they may have as officer of a photo club
+    fileprivate func cleanRoles(memberRolesAndStatus: MemberRolesAndStatus,
+                                photographer: Photographer) -> MemberRolesAndStatus {
+        guard memberRolesAndStatus.isDeceased() == true else { return memberRolesAndStatus }
+
+        var possiblyCorrectedValue = memberRolesAndStatus
+        for eachRole in MemberRole.allCases where memberRolesAndStatus.role[eachRole] == true {
+            possiblyCorrectedValue.role[eachRole] = false
+            ifDebugFatalError("""
+                              Clearing role of \(eachRole) for deceased photographer \(photographer.fullNameFirstLast)
+                              """)
+        }
+        return possiblyCorrectedValue
     }
 
 }