+++ Photo Club Hub/Documentation/ReleaseNotes.md
@@ -7,8 +7,9 @@ Preferences screen
 
 Code Maintenance
 * Tested on iOS 18.2.1, and Xcode 16.2 and MacOS 15.2
-* Extended data model for keyword feature (and updated diagram in README.md on GitHub)
+* Extended data model for keyword feature (and updated diagram in README.md on GitHub).
 * Added documentation in README.md for keyword feature in Level 2 files.
+* Automatically cleared any officer roles for a deceased member. Fatal error in debug version only.
 
 Data Maintenance
 * Added AJ to fcBellusImago